Feat: Implement slash commands for creating new blocks
Commands are implemented in such a way, that it is easy to add new commands, even for (potential) third parties.
Currently only node-creation is supported with slash commands. Commands are looked up by their name and aliases or a custom regex.
A list of default commands can be found in defaultCommands.ts.

A dropdown menu with command suggestions is yet to be implemented.