Feat: Adds hyperlink functionality
Currently uses the link Mark from TipTap. The behavior is to catch any
full links(starting with http) from the pastebin and creating a <a>.
When pasting while text is higlighted, it will make a link out of the
highlighted text. Future implementations will likely change behavior and
implement linking in the inline-menu

Closes #6