WIP: Experimental implementation
This is an attempt to be able to have a url in the textfield persist when selecting text that already is a link. InlineMenu creates a ref, that it passes down when creating a LinkForm. This ref is then called when Link is clicked in the inline menu, which toggles whether the linkform menu is expanded or not. Based on this, the text field is emptied, filled with existing link, or open for input. For this the inline menu had to be converted into a function component. This implementation is implemented very poorly, because of complexities with types and the fact that you can still press outside of the LinkForm window and close it while it still thinks that the menu is expanded. I will either restore to previous commit, or fix this implementation to make it usable.