feat: selecting link shows its url in text field
If the selected text is completely within a marked link, the link menu will now
fill the text field with the matching url so it can be edited. If the
matching mark is still larger than the selected portion, it will be
conserved, so only the selected portion will be overwritten by whatever
is edited. If text is selected that exceeds the marked link, the text
field will show empty, but will still properly create a link on the
entire selection.