Hyperlink integration (#113)
* Feat: Adds hyperlink functionality

Currently uses the link Mark from TipTap. The behavior is to catch any
full links(starting with http) from the pastebin and creating a <a>.
When pasting while text is higlighted, it will make a link out of the
highlighted text. Future implementations will likely change behavior and
implement linking in the inline-menu

Closes #6

* WIP: add menu/ui for links

These interfaces are not yet functional

* WIP: add hooks to LinkForm and change structure

* WIP: Add matching css and react-icon

* WIP: Add link mark functionality to link ui

The barebones functionality now works, and the text field gets cleared
after creating a link. The actual href that gets created however is not
always properly interpreted, so for now you have to provide a full link
with https://*. This will be the next commit.

* WIP: Add custom Mark and clear field on menu exit

Added a custom mark(CustomLink) that extends the Link mark for future
extension. Also added another hook to be able to manipulate what the
link field shows when editor changes. It now clears the text field on
menu exit, so the last used link does not remain.

* WIP: Experimental implementation

This is an attempt to be able to have a url in the textfield persist when selecting text that already is a link. InlineMenu creates a ref, that it passes down when creating a LinkForm. This ref is then called when Link is clicked in the inline menu, which toggles whether the linkform menu is expanded or not. Based on this, the text field is emptied, filled with existing link, or open for input. For this the inline menu had to be converted into a function component. This implementation is implemented very poorly, because of complexities with types and the fact that you can still press outside of the LinkForm window and close it while it still thinks that the menu is expanded. I will either restore to previous commit, or fix this implementation to make it usable.

* Revert "WIP: Experimental implementation"

This reverts commit c41d2987f68f5cee27ebda64f982820ff2364b8a. Postpone
implementation of url showing in the link text field when highlighting an
existing link

* feat: selecting link shows its url in text field

If the selected text is completely within a marked link, the link menu will now
fill the text field with the matching url so it can be edited. If the
matching mark is still larger than the selected portion, it will be
conserved, so only the selected portion will be overwritten by whatever
is edited. If text is selected that exceeds the marked link, the text
field will show empty, but will still properly create a link on the
entire selection.

* Empty "submit" deletes link and code restructure

The entire selection of text will be rid of its link mark when
submitting an empty url. The link button is now uniformly integrated in
InlineMenu and makes use of its types so not much duplicate code
remains.

* Restructure code and add comments

Tooltip is now a component, and the link menu button is not handled
within one function anymore, but there is a dedicated LinkInlineMenuButton
component. css file turned into modules.

* Make all links absolute paths, add CustomLink mark

href's are prepended with "//" so the browser opens them as absolute
paths, meaning that whatever is submitted with the link text field will
be opened in a new tab when the link is pressed. Also added a CustomLink
mark to be able to extend from it.

* Finish styling on link menu

looks beautiful now

* WIP: experiment with auto-focus with hooks and ref

Might revert from this commit considering another approach might be
better. the LinkForm only gets rendered once, so it makes focusing on
the text field after expanding the menu as straight forward.

* Revert "WIP: experiment with auto-focus with hooks and ref"

This reverts commit 37fc93d3cf0dec31dee23005072da5e13c402c7a. Reverted
to still keep implementation in the tree to be able to reference back if
needed. For now, switching the focus to the input field dynamically
seems to be more difficult and involves refs from InlineMenu.

* Move openOnClick function from link to CustomLink

* Apply PR suggestions, comment, further polish

* Update generated files after merge from dev

* Deletes unused CustomLink.tsx and now uses Link

* Add placeholder styling for form for visability

* Structure disabled code, links now integrated.

The link menu does not use Atlaskit components yet. See PR comment for
more detailed information.

Co-authored-by: Yousef El-Dardiry <yousefdardiry@gmail.com>