Change PageLayout into manual wrapper/layout
The components rendered by Login and Registration are cleaner this way
as it only uses the manual wrappers now, which used to be wrappers that
Matrix used. All the content is passed in Login/Registration, so the
components like AuthPage, AuthHeader, etc. are ONLY wrappers with
styling. This styling is soon to be moved to Component.module.css(which
may change this structure again)