change login default field values, see desc
There were some inconsistencies with states and assigning them to
defaul values. For login, Matrix stores the state of the username,
phoneCountry and phoneNumber fields in Login.tsx, and password in
PasswordLogin.tsx. for registration it stores all the fields in
RegistrationForm.tsx. The field states were all removed, but I restored
the ones in Login.tsx, as that state also gets affected by submission,
and the values of that state will now be assigned to the PasswordLogin
form that gets created.