Implement follow up TODO's for authentication frontend (#167)
* Change PageLayout into manual wrapper/layout

The components rendered by Login and Registration are cleaner this way
as it only uses the manual wrappers now, which used to be wrappers that
Matrix used. All the content is passed in Login/Registration, so the
components like AuthPage, AuthHeader, etc. are ONLY wrappers with
styling. This styling is soon to be moved to Component.module.css(which
may change this structure again)

* remove ui comonents and put styling in modules

See PR #167 for details on this structure desicion

* Add restructure from Login to Registration too

* change login default field values, see desc

There were some inconsistencies with states and assigning them to
defaul values. For login, Matrix stores the state of the username,
phoneCountry and phoneNumber fields in Login.tsx, and password in
PasswordLogin.tsx. for registration it stores all the fields in
RegistrationForm.tsx. The field states were all removed, but I restored
the ones in Login.tsx, as that state also gets affected by submission,
and the values of that state will now be assigned to the PasswordLogin
form that gets created.

* add documentation to Login and PasswordLogin

* Add comments, change a lot of structure and more

* organize styling/other code

* auth ui improvements and sso support

Co-authored-by: Yousef El-Dardiry <yousefdardiry@gmail.com>